Moved buffermodel.ts to webgl/
