+++ src/client/scripts/esm/game/rendering/arrows/arrowlegalmovehighlights.ts
@@ -8,7 +8,7 @@
 
 
 import type { Color } from "../../../../../../shared/util/math/math.js";
-import type { BufferModelInstanced } from "../buffermodel.js";
+import type { BufferModelInstanced } from "../../../webgl/buffermodel.js";
 import type { LegalMoves } from "../../../../../../shared/chess/logic/legalmoves.js";
 import type { Vec3 } from "../../../../../../shared/util/math/vectors.js";
 import type { Piece } from "../../../../../../shared/chess/util/boardutil.js";
