+++ src/client/scripts/esm/game/rendering/highlights/squarerendering.ts
@@ -14,7 +14,7 @@
 import space from "../../misc/space.js";
 import instancedshapes from "../instancedshapes.js";
 import bd from "../../../../../../shared/util/bigdecimal/bigdecimal.js";
-import { BufferModelInstanced, createModel_Instanced } from "../buffermodel.js";
+import { BufferModelInstanced, createModel_Instanced } from "../../../webgl/buffermodel.js";
 
 import type { Coords } from "../../../../../../shared/chess/util/coordutil.js";
 import type { Color } from "../../../../../../shared/util/math/math.js";
