+++ src/client/scripts/esm/webgl/buffermodel.ts
@@ -11,14 +11,14 @@
  */
 
 
-import type { Vec3 } from '../../../../../shared/util/math/vectors.js';
+import type { Vec3 } from '../../../../shared/util/math/vectors.js';
 
-import { ShaderProgram } from '../../webgl/ShaderProgram.js';
-import { Attributes_All, ProgramManager, ProgramMap } from '../../webgl/ProgramManager.js';
-import { createBufferFromData, updateBufferIndices } from './buffers.js';
-import camera, { Mat4 } from './camera.js';
+import { ShaderProgram } from './ShaderProgram.js';
+import { Attributes_All, ProgramManager, ProgramMap } from './ProgramManager.js';
+import { createBufferFromData, updateBufferIndices } from '../game/rendering/buffers.js';
+import camera, { Mat4 } from '../game/rendering/camera.js';
 // @ts-ignore
-import mat4 from './gl-matrix.js';
+import mat4 from '../game/rendering/gl-matrix.js';
 
 
 // Type Definitions -----------------------------------------------------------------------
